fix(sonar): deduplicate env parsing in configs and scripts
