+++ apps/kernel/server.js
@@ -1,29 +1,7 @@
 // Load .env.local (Node doesn't do this automatically like Next.js does)
-const { readFileSync, existsSync } = require('fs');
 const envPath = require('path').join(__dirname, '.env.local');
-
-function parseEnvLine(line) {
-  const trimmed = line.trim();
-  if (!trimmed || trimmed.startsWith('#')) return null;
-  const eqIdx = trimmed.indexOf('=');
-  if (eqIdx <= 0) return null;
-  const key = trimmed.slice(0, eqIdx).trim();
-  if (!/^[A-Z_][A-Z0-9_]*$/.test(key)) return null;
-  let value = trimmed.slice(eqIdx + 1).trim();
-  if (
-    (value.startsWith('"') && value.endsWith('"')) ||
-    (value.startsWith("'") && value.endsWith("'"))
-  ) {
-    value = value.slice(1, -1);
-  }
-  return { key, value };
-}
-if (existsSync(envPath)) {
-  for (const line of readFileSync(envPath, 'utf-8').split('\n')) {
-    const parsed = parseEnvLine(line);
-    if (parsed && !process.env[parsed.key]) process.env[parsed.key] = parsed.value;
-  }
-}
+const envUtils = require('../../scripts/env-utils.js');
+envUtils.loadEnvFileIntoProcessEnv(envPath);
 
 const { createServer } = require('http');
 const { parse } = require('url');
